src/loader.py
import logging

"""
Module for loading models and data.
"""

# Configure logging
logger = logging.getLogger(__name__)


class DataLoader:
    """
    Professional Data Loader class.
    """
    def __init__(self):
        """
        Initialize the data loader and check for optimum-intel.
        """
        try:
            from optimum.intel import OVModelForSpeechSeq2Seq, OVModelForSeq2SeqLM  # noqa: F401
            logger.info("Optimum Intel libraries loaded")
        except ImportError:
            logger.warning("Optimum Intel not available")

    def load_data(self):
        """
        Simulation of data loading.
        """
        logger.info("Loading data...")
        return {"samples": []}
